refactor(core): optimize repeated code for swipe directions

+++ core/embed/rust/src/ui/model_mercury/component/swipe_up_screen.rs
@@ -49,7 +49,7 @@ impl<T: Swipable + Component> Component for SwipeUpScreen<T> {
                 return Some(SwipeUpScreenMsg::Swiped);
             }
             Some(SwipeDetectMsg::Move(dir, progress)) => {
-                Event::Swipe(SwipeEvent::Move(dir, progress))
+                Event::Swipe(SwipeEvent::Move(dir, progress as i16))
             }
             _ => event,
         };
